requirements.txt
appdirs==1.4.3
autopep8==1.2.2
boto3==1.4.4
botocore==1.5.48
coverage==4.0.3
Django==1.8.2
django-cors-headers==1.1.0
django-enumfield==1.2.1
django-extensions==1.5.5
django-extras==0.3
django-filter==0.10.0
django-nose==1.4.3
django-redis==4.4.3
django-rest-swagger==0.3.2
django-storages==1.5.2
djangorestframework==3.2.2
docutils==0.13.1
drf-nested-routers==0.9.0
elasticsearch==2.4.0
factory-boy==2.5.2
fake-factory==0.5.1
flake8==2.5.4
gunicorn==19.3.0
jmespath==0.9.2
Markdown==2.6.2
mccabe==0.4.0
mockredispy==2.9.3
msgpack-python==0.4.7
nose==1.3.7
packaging==16.8
pep8==1.7.0
Pillow==2.9.0
psycopg2==2.6.1
pyflakes==1.0.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2015.4
PyYAML==3.11
redis==2.10.5
requests==2.7.0
s3transfer==0.1.10
six==1.9.0
urllib3==1.21.1
wheel==0.24.0
